requirements.txt
mysql.connector
pymysql
cryptography
mysql-connector-python
